refactor: move workshop and panels progress view
